Move plans by week logic into sessions context
The plans context seems unnecessary and makes it more difficult to work
with the sessions.

Sessions with associations are preloaded in db queries, then
inactive recipients are filtered out later, instead of trying to do
this in ecto.

This makes it a bit easier to work with, even if the queries are a bit
less efficient.

Still need to figure out what to do with duplicated recipient status
logic.